better error handling in mkwin/packageexe.sh

+++ mkwin/packageexe.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
+
+#example: mkwin/packageexe.sh $PWD 1.27
 ROOTDIR=$1
 WINDIR=$ROOTDIR/mkwin
 VERSION=$2
 SCALA_VERSION=2.13
 JRE=jre11
-echo "ROOTDIR $ROOTDIR WINDIR $WINDIR VERSION $VERSION"
+echo "ROOTDIR $ROOTDIR"
+echo "WINDIR $WINDIR"
+echo "VERSION $VERSION"
 cp $WINDIR/launch4j_config.mod $WINDIR/launch4j_config.xml
 JARFILE="$ROOTDIR/prjwin/target/scala-$SCALA_VERSION/dbtarzan-assembly-$VERSION.jar"
 OUTFILE="$WINDIR/dbtarzan_$VERSION.exe" 
@@ -12,12 +16,20 @@ ICONFILE="$WINDIR/monkey-face-cartoon.ico"
 JARESCAPED="${JARFILE//\//\\\/}"
 OUTESCAPED="${OUTFILE//\//\\\/}"
 ICONESCAPED="${ICONFILE//\//\\\/}"
-sed -i "s/JARFILE/$JARESCAPED/g" $WINDIR/launch4j_config.xml
-sed -i "s/OUTFILE/$OUTESCAPED/g" $WINDIR/launch4j_config.xml
-sed -i "s/ICONFILE/$ICONESCAPED/g" $WINDIR/launch4j_config.xml
-$ROOTDIR/../../bin/launch4j/launch4j $WINDIR/launch4j_config.xml
-rm -r $JRE
-wget -O $JRE.zip "https://api.adoptopenjdk.net/v2/binary/nightly/openjdk11?openjdk_impl=hotspot&os=windows&arch=x64&release=latest&type=jre"
-unzip $JRE.zip
-mv jdk* $JRE
-makensis -DVERSION=$VERSION $WINDIR/nsis.nsi
+LAUNCH4J_CONFIG=$WINDIR/launch4j_config.xml
+echo "LAUNCH4J_CONFIG $LAUNCH4J_CONFIG"
+sed -i "s/JARFILE/$JARESCAPED/g" $LAUNCH4J_CONFIG
+sed -i "s/OUTFILE/$OUTESCAPED/g" $LAUNCH4J_CONFIG
+sed -i "s/ICONFILE/$ICONESCAPED/g" $LAUNCH4J_CONFIG
+#it must be in unix format to be able to work. dos2unix can help
+$ROOTDIR/../../bin/launch4j/launch4j $LAUNCH4J_CONFIG
+if [ $? -eq 0 ]
+then
+  rm -r $JRE
+  wget -O $JRE.zip "https://github.com/adoptium/temurin11-binaries/releases/download/jdk-11.0.19%2B7/OpenJDK11U-jre_x64_windows_hotspot_11.0.19_7.zip"
+  unzip $JRE.zip
+  mv jdk* $JRE
+  makensis -DVERSION=$VERSION $WINDIR/nsis.nsi
+else
+  echo "launch4j failed with exit code $?"
+fi
